Refactor recommended nodes
