+++ explorer-v2/src/app/(pages)/table/page.tsx
@@ -7,11 +7,9 @@ import { Wrapper } from "@/components/wrapper";
 import { Box, Stack } from "@mui/material";
 // import Grid from "@mui/material/Grid2";
 
-import { RECOMMENDED_NODES } from "@/app/constants"; // ⬅ dynamic Promise<number[]>
 
 export default async function ExplorerPage() {
   // Resolve once on the server and pass IDs to client components
-  const recommendedIds = await RECOMMENDED_NODES;
 
   return (
     <ContentLayout>
@@ -21,7 +19,7 @@ export default async function ExplorerPage() {
           <NodeAndAddressSearch />
         </Stack>
         <Box sx={{ mt: 5 }}>
-          <NodeTableWithAction recommendedIds={recommendedIds} />
+          <NodeTableWithAction />
         </Box>
         {/* <Grid container columnSpacing={5} rowSpacing={5} mt={10}>
           <Grid size={12}>
