+++ explorer-v2/src/app/constants.ts
@@ -1,5 +1,3 @@
 export const TABLET_WIDTH = "(min-width:700px)";
 
-import { getRecommendedNodes } from "./lib/recommended";
 
-export const RECOMMENDED_NODES = getRecommendedNodes();
